requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
asyncpg==0.29.0
asyncio-rcon==0.2.2
aioftp==0.22.3
aiohttp==3.9.3
